requirements.txt
tensorflow
keras
pandas
numpy
scipy
scikit-image
scikit-learn
h5py
tqdm
seaborn
matplotlib
colorama
mongoengine
click
pyyaml
